Refactor CLI structure by moving main logic to main.py and updating entry point
